+++ src/lvmguider/__init__.py
@@ -1,8 +1,14 @@
 # encoding: utf-8
 
 import pathlib
+import warnings
+
+from astropy.wcs import FITSFixedWarning
 
 from sdsstools import Configuration, get_logger, get_package_version
+
+
+warnings.simplefilter("ignore", category=FITSFixedWarning)
 
 
 NAME = "lvmguider"
