+++ src/lvmguider/etc/lvmguider.yml
@@ -12,6 +12,9 @@ pid:
 max_correction: 3600
 guide_tolerance: 3
 
+xz_full_frame: [2500.0, 1000.0]
+xz_ag_frame: [800.0, 550.0]
+
 offset:
   axis_error: 1.0
   timeout:
